+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Links in Group Post/Reply do not work [#534](https://github.com/rokwire/illinois-app/issues/534).
 ### Changed
 - "Reply" and "Delete" buttons order for post a reply [#532](https://github.com/rokwire/illinois-app/issues/532).